@@ -5,42 +5,45 @@
   "type": "module",
   "scripts": {
     "test": "vitest",
-    "dev": "vite --port 6001",
+    "dev": "vite --port 5013",
     "build": "tsc -b && vite build",
+    "deploy": "scp -r dist andrew@cofasv32:~/apps/builds/step-up-vite-ts",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@azure/msal-browser": "^4.9.1",
-    "@azure/msal-react": "^3.0.8",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "@azure/msal-browser": "^4.24.0",
+    "@azure/msal-react": "^3.0.20",
+    "@tanstack/react-query": "^5.62.12",
+    "motion": "^12.23.24",
+    "react": "^19.1.1",
+    "react-dom": "^19.0.0",
+    "react-error-boundary": "^6.0.0",
     "react-hook-form": "^7.53.0",
-    "react-query": "^3.39.3",
-    "react-router-dom": "^6.25.1",
+    "react-router": "^7.9.3",
     "react-toastify": "^10.0.5"
   },
   "devDependencies": {
+    "@eslint/js": "^9.38.0",
+    "@tailwindcss/vite": "^4.1.16",
+    "@tanstack/react-query-devtools": "^5.90.2",
     "@testing-library/jest-dom": "^6.4.8",
-    "@testing-library/react": "^16.0.0",
-    "@types/jest": "^29.5.12",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@types/react-test-renderer": "^18.3.0",
+    "@testing-library/react": "^16.3.0",
+    "@types/node": "^24.9.1",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
     "@typescript-eslint/eslint-plugin": "^7.15.0",
     "@typescript-eslint/parser": "^7.15.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "autoprefixer": "^10.4.19",
-    "daisyui": "^4.12.10",
+    "@vitejs/plugin-react": "^4.3.2",
+    "daisyui": "^5.3.10",
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.7",
     "jsdom": "^24.1.1",
-    "postcss": "^8.4.39",
-    "tailwindcss": "^3.4.6",
-    "ts-mockito": "^2.6.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.4",
+    "tailwindcss": "^4.1.14",
+    "typescript": "^5.9.3",
+    "vite": "^7.1.7",
+    "vite-tsconfig-paths": "^5.1.4",
     "vitest": "^2.0.5"
   }
 }
