@@ -1,9 +1,10 @@
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom"
-import { QueryClient, QueryClientProvider } from "react-query"
-import { ReactQueryDevtools } from 'react-query/devtools'
+import { BrowserRouter as Router, Route, Routes } from "react-router"
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { ToastContainer } from "react-toastify"
-import { AppProvider } from "./context/App/AppContext"
 import { APP_BASE } from './config'
+import { StepUpProvider } from "./components/step-up/context"
+import { RosterProvider } from "./components/roster/context"
 import 'react-toastify/dist/ReactToastify.css'
 
 // Components
@@ -17,20 +18,26 @@ const queryClient = new QueryClient()
 function App() {
   
   return (
-    <AppProvider>
-      <QueryClientProvider client={queryClient}>
-        <Router basename={APP_BASE} future={{ v7_startTransition: true }}>
-          <Routes>
-            <Route path="/" element={<Login />} />
-            <Route path="/home" element={<Home />} />
-            <Route path="/rosters" element={<Rosters />} />
-            <Route path="/*" element={<Redirect />} />
-          </Routes>
-        </Router>
-        <ReactQueryDevtools initialIsOpen={false} />
-        <ToastContainer />
-      </QueryClientProvider>
-    </AppProvider>
+    <QueryClientProvider client={queryClient}>
+      <Router basename={APP_BASE}>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/home" element={
+            <StepUpProvider>
+              <Home />
+            </StepUpProvider>
+          } />
+          <Route path="/rosters" element={
+            <RosterProvider>
+              <Rosters />
+            </RosterProvider>
+          } />
+          <Route path="/*" element={<Redirect />} />
+        </Routes>
+      </Router>
+      <ReactQueryDevtools initialIsOpen={false} />
+      <ToastContainer />
+    </QueryClientProvider>
   )
 }
 
