@@ -1,21 +1,22 @@
-import { Link } from 'react-router-dom'
-import { APP_TITLE } from '../../../config'
-import styles from './Header.module.css'
+import { useSetActivePage } from './hooks'
 
 // Components
 import * as Components from './components'
 
 function Header() {
+  useSetActivePage()
 
   return (
-    <header data-testid="header" className={styles.header}>
-      <Link to={'/'}>
-        <div className={styles.title}>
-          <h1 className={styles.h1}>Franklin Fire Department</h1>
-          <h2 className={styles.h2}>{APP_TITLE}</h2>
+    <header className="flex flex-col font-[play] w-full">
+      <div className="flex justify-between font-[play] tracking-[.25rem] items-center bg-gradient-to-r from-primary/80 to-primary py-1 px-8 h-[15vh] shadow-xl 2xl:h-[10vh]">
+        <Components.Title />
+
+        <div className="flex gap-2">
+          <Components.Buttons />
         </div>
-      </Link>
-      <Components.Buttons />
+      </div>
+
+      <Components.HomeLink />
     </header>
   )
 }
