@@ -1,3 +1,6 @@
-export const APP_TITLE="Step Up Pay and Rosters"
+export const APP_TITLE="FFD Step Up Pay & Rosters"
 export const APP_BASE="/step-up"
-export const API_URL="https://dev.franklintn.gov/api/v2/ffd/step-up"
+// export const API_URL="https://cofasv38.franklin-gov.com/api/v2/ffd/step-up"
+export const API_URL="https://fireapps.franklintn.gov/api/v2/ffd/step-up"
+export const NODE_ENV: 'development' | 'test' | 'production' = 'production'
+export const CLIENT_ID = 'beec6da4-ba46-45c7-9630-ddb7df68b3da'
