@@ -1,11 +1,11 @@
 import { API_URL as baseUrl } from "../../config"
 
 // Types
-import { GetStaffResponse, GetRosterResponse } from './types'
+import * as AppTypes from '@/context/App/types'
 
 // Get all staff
 // GET /api/v1/ffd/step-up/staff
-export const getStaff = async (headers: Headers): Promise<GetStaffResponse> => {
+export const getStaff = async (headers: Headers): Promise<AppTypes.ServerResponse & { data: AppTypes.StaffInterface[] }> => {
   const res = await fetch(`${ baseUrl }/staff`, { headers })
 
   return await res.json()
@@ -13,7 +13,7 @@ export const getStaff = async (headers: Headers): Promise<GetStaffResponse> => {
 
 // Get roster
 // GET /api/v1/ffd/step-up/roster?date=
-export const getRoster = async (date: string, headers: Headers): Promise<GetRosterResponse> => {
+export const getRoster = async (date: string, headers: Headers): Promise<AppTypes.ServerResponse & { data: AppTypes.RosterEntryInterface[] }> => {
   const res = await fetch(`${ baseUrl }/rosters?date=${ date }`, { headers })
 
   return await res.json()
