@@ -1,15 +1,16 @@
-import { LogLevel } from '@azure/msal-browser';
+import { LogLevel } from '@azure/msal-browser'
+import { CLIENT_ID } from '@/config'
 
 export const msalConfig = {
     auth: {
-        clientId: 'beec6da4-ba46-45c7-9630-ddb7df68b3da',
+        clientId: CLIENT_ID,
         authority: 'https://login.microsoftonline.com/f6644f52-f834-4a2f-a433-e6bc40d7c17f/',
-        redirectUri: 'https://dev.franklintn.gov/step-up',
-        postLogoutRedirectUri: '/',
+        redirectUri: 'https://fireapps.franklintn.gov/step-up',
+        postLogoutRedirectUri: 'https://fireapps.franklintn.gov/',
         navigateToLoginRequestUrl: false
     },
     cache: {
-        cacheLocation: 'sessionStorage', 
+        cacheLocation: 'localStorage',
         storeAuthStateInCookie: false
     },
     system: {
@@ -38,5 +39,5 @@ export const msalConfig = {
 
 export const loginRequest = {
     scopes: ["openid", "profile"],
-    redirectUri: "https://dev.franklintn.gov/step-up"
+    redirectUri: "https://fireapps.franklintn.gov/step-up"
 };
