@@ -1,26 +1,53 @@
 import { PublicClientApplication } from "@azure/msal-browser"
 import { MsalProvider } from "@azure/msal-react"
+import { useEffect, useState } from "react"
+import { NODE_ENV } from "@/config"
 import { msalConfig } from "../config"
 
 // Types
-import { ReactNode } from "react"
 import { AuthenticationResult, EventType } from "@azure/msal-browser"
 
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const msalInstance = new PublicClientApplication(msalConfig)
+export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+  const [msalInstance, setMsalInstance] = useState<PublicClientApplication | null>(null)
+  const [isInitialized, setIsInitialized] = useState(false)
 
-  if(!msalInstance.getActiveAccount() && msalInstance.getAllAccounts().length > 0) {
-    msalInstance.setActiveAccount(msalInstance.getAllAccounts()[0])
-  }
-  
-  msalInstance.addEventCallback((event) => {
-    const authenticationResult = event.payload as AuthenticationResult
-    const account = authenticationResult?.account
+  if(NODE_ENV === 'development') return (
+    <>
+      {children}
+    </>
+  )
 
-    if(event.eventType === EventType.LOGIN_SUCCESS && account) {
-      msalInstance.setActiveAccount(account)
+  useEffect(() => {
+    const initializeMsal = async () => {
+      const instance = new PublicClientApplication(msalConfig)
+      
+      await instance.initialize()
+      
+      const accounts = instance.getAllAccounts()
+      
+      if(!instance.getActiveAccount() && accounts.length > 0) {
+        instance.setActiveAccount(accounts[0])
+      }
+      
+      instance.addEventCallback((event) => {
+        const authenticationResult = event.payload as AuthenticationResult
+        const account = authenticationResult?.account
+
+        if(event.eventType === EventType.LOGIN_SUCCESS && account) {
+          instance.setActiveAccount(account)
+        }
+      })
+
+      setMsalInstance(instance)
+      setIsInitialized(true)
     }
-  })
+
+    initializeMsal()
+  }, [])
+
+  if(!isInitialized || !msalInstance) {
+    return <div>Initializing authentication...</div>
+  }
 
   return <MsalProvider instance={msalInstance}>
     {children}
