@@ -1,20 +1,20 @@
-import { useRedirect } from '../Redirect/hooks'
+import { useRedirectAfterLogin } from '@/helpers/hooks'
 import { useGetStaff } from './hooks'
 
 // Components
 import Layout from '../../components/layout/Layout'
-import TableContainer from '../../components/table/TableContainer'
+import TableContainer from '../../components/step-up/containers/TableContainer'
 import HandleLoading from '../../utils/HandleLoading'
 
 function Home() {
-  useRedirect()
+  useRedirectAfterLogin()
 
   const { data, isSuccess } = useGetStaff()
 
   return (
     <Layout>
       <HandleLoading isLoaded={isSuccess}>
-        <TableContainer staff={data?.data || []} />
+        <TableContainer staff={data?.data} />
       </HandleLoading>
     </Layout>
   )
