@@ -1,4 +1,4 @@
-import { useRedirect } from "./hooks"
+import { useRedirect } from './hooks'
 
 function Redirect() {
   useRedirect()
