@@ -1,21 +1,19 @@
 import { useContext } from 'react'
-import { useQuery } from 'react-query'
-import AppContext from '../../context/App/AppContext'
-import { useGetToken, useEnableQuery } from '../../helpers/hooks'
-import { authHeaders } from '../../helpers/utils'
-import { getRoster } from '../../context/App/AppActions'
+import { useQuery } from '@tanstack/react-query'
+import { useEnableQuery } from '@/helpers/hooks'
+import { authHeaders } from '@/helpers/utils'
+import * as AppActions from '@/context/App/AppActions'
+import RosterCtx from '@/components/roster/context'
 
 export const useGetRoster = () => { // Get roster from Telestaff API
-  const { date } = useContext(AppContext)
+  const { rosterDate } = useContext(RosterCtx)
 
-  const token = useGetToken()
-
-  const enabled = useEnableQuery(token)
+  const { enabled, token } = useEnableQuery()
 
   const today = new Date()
   const todayStr = `${ today.getFullYear() }-${ today.getMonth() + 1 }-${ today.getDate() }`
 
-  const queryDate = date || todayStr
+  const queryDate = rosterDate || todayStr
 
-  return useQuery(['roster', queryDate], () => getRoster(queryDate, authHeaders(token)), { enabled })
+  return useQuery({ queryKey: ['getRosters', queryDate], queryFn: () => AppActions.getRoster(queryDate, authHeaders(token)), enabled: enabled && !!token && !!queryDate, staleTime: Infinity })
 }
