@@ -1,20 +1,20 @@
-import { useRedirect } from "../Redirect/hooks"
+import { useRedirectAfterLogin } from "@/helpers/hooks"
 import { useGetRoster } from "./hooks"
 
 // Components
-import Layout from "../../components/layout/Layout"
-import RosterContainer from "../../components/roster/RosterContainer"
-import HandleLoading from "../../utils/HandleLoading"
+import Layout from "@/components/layout/Layout"
+import RosterContainer from "@/components/roster/containers/RosterContainer"
+import HandleLoading from "@/utils/HandleLoading"
 
 function Rosters() {
-  useRedirect()
-  
+  useRedirectAfterLogin()
+
   const { data, isSuccess } = useGetRoster()
 
   return (
     <Layout>
       <HandleLoading isLoaded={isSuccess}>
-        <RosterContainer rosters={data?.data || []} />
+        <RosterContainer rosters={data?.data} />
       </HandleLoading>
     </Layout>
   )
