@@ -1,18 +1,25 @@
 import { defineConfig } from 'vite'
-import { APP_BASE } from './src/config'
 import react from '@vitejs/plugin-react'
-import { UserConfig } from 'vitest/config'
+import path from 'path'
+import { fileURLToPath } from 'url'
+import tsconfigPaths from "vite-tsconfig-paths"
+import tailwindcss from '@tailwindcss/vite'
 
-/// <reference types="vitest" />
-/// <reference types="vite/client" />
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
-const config: UserConfig = {
-  plugins: [
-    react(),
-  ],
-  base: APP_BASE,
+export default defineConfig({
+  plugins: [tailwindcss(), react(), tsconfigPaths()],
+  base: '/step-up',
   resolve: {
     alias: {
+      '@': path.resolve(__dirname, './src'),
+      '@/components': path.resolve(__dirname, './src/components'),
+      '@/config': path.resolve(__dirname, './src/config'),
+      '@/hooks': path.resolve(__dirname, './src/hooks'),
+      '@/pages': path.resolve(__dirname, './src/pages'),
+      '@/utils': path.resolve(__dirname, './src/utils'),
+      '@/assets': path.resolve(__dirname, './src/assets'),
     },
   },
   optimizeDeps: {
@@ -23,12 +30,7 @@ const config: UserConfig = {
         }
     }
   },
-  test: {
-    globals: true,
-    environment: 'jsdom',
-    css: false,
-    setupFiles: "./src/test/setup.ts"
+  server: {
+    allowedHosts: ['cofasv38.franklin-gov.com', 'istest.franklintn.gov', 'dev.franklintn.gov']
   }
-}
-
-export default defineConfig(config)
+})
